database-migration-py/requirements.txt
# ================================================================
# Financer Database Migration System - Python Dependencies
# ================================================================

# Database Drivers
psycopg2-binary==2.9.9      # PostgreSQL adapter
pymongo==4.6.0               # MongoDB driver

# HTTP Client for Config Server
requests==2.31.0             # HTTP library
requests-auth==7.0.0         # Authentication support

# Data Handling
pydantic==2.5.0              # Data validation and parsing
python-dotenv==1.0.0         # Environment variables

# Logging and Utilities
colorlog==6.8.0              # Colored logging
rich==13.7.0                 # Rich text and beautiful formatting
click==8.1.7                 # Command line interface

# Date/Time handling
python-dateutil==2.8.2      # Date utilities

# JSON handling
ujson==5.9.0                 # Fast JSON parsing

# Testing (optional)
pytest==7.4.3               # Testing framework
pytest-cov==4.1.0           # Coverage testing